mac-butler/requirements.txt
faster-whisper
sounddevice
numpy
pyyaml
requests
pynput
SpeechRecognition
psutil
